requirements.txt
# Core LangChain (LOCKED)
langchain==0.2.14
langchain-core==0.2.43
langchain-community==0.2.12
langchain-openai==0.1.23
langchain-text-splitters==0.2.4
langserve==0.2.3

# APIs
openai>=1.40.0
groq==0.37.1
langchain-groq==0.1.10

# Web / app
fastapi==0.128.0
uvicorn==0.40.0
sse-starlette==3.1.1
streamlit

# Docs / parsing
python-dotenv
beautifulsoup4
bs4
pypdfw
PyPDF2
wikipedia
arxiv

# Vector / ML
chromadb==0.5.5
faiss-cpu
sentence-transformers
torch
scikit-learn
scipy

# Optional
langchainhub
cassio
